# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Waimai CPS API
  version: 1.0.0
servers:
  - url: https://api.example.com/api/v1
paths:
...
  /mini/home:
    get: {summary: 首页内容, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/profile:
    get: {summary: 用户资料, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
    put: {summary: 更新用户资料, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/orders:
    get: {summary: 用户订单, security: [{miniBearer: []}], parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/wallet:
    get: {summary: 钱包汇总, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/wallet/entries:
    get: {summary: 钱包流水, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/invitations:
    get: {summary: 直接邀请记录, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
  /mini/promotion-links:
    post: {summary: 生成官方联盟推广链接, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}, "400": {$ref: "#/components/responses/Error"}}}
  /mini/withdrawals:
    get: {summary: 提现列表, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
    post:
      summary: 提交提现并冻结余额
      security: [{miniBearer: []}]
      parameters: [{name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 128}}]
      responses: {"200": {$ref: "#/components/responses/Success"}, "400": {$ref: "#/components/responses/Error"}}
  /mini/withdrawals/{id}/cancel:
    post:
      summary: 取消本人待审核提现
      security: [{miniBearer: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {"200": {$ref: "#/components/responses/Success"}}
  /mini/share-scenes:
    post: {summary: 生成 24 小时有效邀请场景参数, security: [{miniBearer: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
...
  /platform/tenants:
    get: {summary: 平台管理员查询租户, security: [{adminCookie: []}], responses: {"200": {$ref: "#/components/responses/Success"}}}
components:
  securitySchemes:
    miniBearer: {type: http, scheme: bearer, bearerFormat: opaque-random-token}
    adminCookie: {type: apiKey, in: cookie, name: ADMIN_SESSION}
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  schemas:
...
    ApiResponse: {type: object, required: [code, message, requestId], properties: {code: {type: string}, message: {type: string}, data: {}, requestId: {type: string, format: uuid}}}
  responses:
    Success: {description: 成功, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}
    Error: {description: 业务错误, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}